desimaltall-barnehageplass * Bruker parsefloat istedenfor parseint når vi konverterer antalltimer til tall. parseInt gir java.lang.NumberFormatException om man sender med et '.'. * La til test for å sjekke at desimaltall funker

+++ src/main/java/no/nav/kontantstotte/innsending/oppsummering/html/mapping/BarnehageplassMapping.java
@@ -59,7 +59,7 @@ public class BarnehageplassMapping extends BolkMapping {
                     );
                     break;
                 case harBarnehageplass:
-                    Element harBarnehageplassAntallTimer = Integer.parseInt(barnehageplass.harBarnehageplassAntallTimer) > 33 ?
+                    Element harBarnehageplassAntallTimer = Float.parseFloat(barnehageplass.harBarnehageplassAntallTimer) > 33.0 ?
                             Element.nyttSvar(
                                     tekster.hentTekst(HAR_BARNEHAGEPLASS_ANTALL_TIMER.getNokkel(), erFlerling),
                                     barnehageplass.harBarnehageplassAntallTimer,
